# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Robot Shop",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "搜索时支持用户名、昵称、手机和邮箱筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户列表",
                "operationId": "ListUsers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "昵称",
                        "name": "nickname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "手机",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.UserSearchResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建一个新的用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "创建用户",
                "operationId": "UserCreate",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "更新用户",
                "operationId": "UserUpdate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除指定ID的用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "删除用户",
                "operationId": "UserDelete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "支持用户名或邮箱登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "登录",
                "operationId": "Login",
                "parameters": [
                    {
                        "description": "登录凭证",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.LoginResponse"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "目前只支持通过邮箱进行注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "注册",
                "operationId": "Register",
                "parameters": [
                    {
                        "description": "注册信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
...
        "/users/me": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取当前用户的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取当前用户",
                "operationId": "GetCurrentUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.UserResponse"
                        }
                    }
                }
            }
        },
        "/users/me/menu": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取当前用户的菜单列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户菜单",
                "operationId": "GetUserMenu",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.MenuSearchResponse"
                        }
                    }
                }
            }
        },
        "/users/me/permission": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取当前用户的权限列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户权限",
                "operationId": "GetUserPermission",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.UserPermissionResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Menu": {
            "type": "object",
            "properties": {
                "component": {
                    "description": "绑定的组件",
                    "type": "string"
                },
                "hideInMenu": {
                    "description": "是否保活",
                    "type": "boolean"
                },
                "icon": {
                    "description": "图标，使用字符串表示",
                    "type": "string"
                },
                "id": {
                    "description": "唯一id，使用整数表示",
                    "type": "integer"
                },
                "keepAlive": {
                    "description": "是否保活",
                    "type": "boolean"
                },
                "locale": {
                    "description": "本地化标识",
                    "type": "string"
                },
                "name": {
                    "description": "同路由中的name，唯一标识",
                    "type": "string"
                },
                "parentId": {
                    "description": "父级菜单的id，使用整数表示",
                    "type": "integer"
                },
                "path": {
                    "description": "地址",
                    "type": "string"
                },
                "redirect": {
                    "description": "重定向地址",
                    "type": "string"
                },
                "title": {
                    "description": "展示名称",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "是否保活",
                    "type": "string"
                },
                "url": {
                    "description": "iframe模式下的跳转url，不能与path重复",
                    "type": "string"
                },
                "weight": {
                    "description": "排序权重",
                    "type": "integer"
                }
            }
        },
        "MenuList": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Menu"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
...
        "User": {
            "type": "object",
            "required": [
                "email",
                "nickname",
                "username"
            ],
            "properties": {
                "avatar": {
                    "description": "头像",
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "example": "zhangsan@example.com"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string",
                    "example": "Jackal"
                },
                "phone": {
                    "description": "手机",
                    "type": "string",
                    "example": "+86-13966668888"
                },
                "roles": {
                    "description": "角色",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        ""
                    ]
                },
                "status": {
                    "description": "状态 0:待激活 1:正常 2:禁用",
                    "type": "integer",
                    "example": 1
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "example": "zhangsan"
                }
            }
        },
        "UserList": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
...
        "backend_api_v1.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "example": "zhangsan"
                }
            }
        },
        "backend_api_v1.LoginResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/backend_api_v1.LoginResponseData"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "backend_api_v1.LoginResponseData": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                }
            }
        },
...
        "backend_api_v1.MenuSearchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/MenuList"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "backend_api_v1.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "example": "zhangsan@example.com"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "backend_api_v1.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "backend_api_v1.UserPermissionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/backend_api_v1.UserPermissionResponseData"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "backend_api_v1.UserPermissionResponseData": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "backend_api_v1.UserRequest": {
            "type": "object",
            "required": [
                "email",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "zhangsan@example.com"
                },
                "nickname": {
                    "type": "string",
                    "example": "Jackal"
                },
                "phone": {
                    "type": "string",
                    "example": "+86-13966668888"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        ""
                    ]
                },
                "status": {
                    "description": "状态 0:待激活 1:正常 2:禁用",
                    "type": "integer",
                    "example": 1
                },
                "username": {
                    "type": "string",
                    "example": "zhangsan"
                }
            }
        },
        "backend_api_v1.UserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/User"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "backend_api_v1.UserSearchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/UserList"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}